+++ src/black/__init__.py
@@ -36,8 +36,13 @@ from click.core import ParameterSource
 from dataclasses import replace
 from mypy_extensions import mypyc_attr
 
-from black.const import DEFAULT_LINE_LENGTH, DEFAULT_INCLUDES, DEFAULT_EXCLUDES
-from black.const import STDIN_PLACEHOLDER
+from black.const import (
+    DEFAULT_LINE_LENGTH,
+    DEFAULT_INCLUDES,
+    DEFAULT_EXCLUDES,
+    STDIN_PLACEHOLDER,
+    LogLevel,
+)
 from black.nodes import STARS, syms, is_simple_decorator_expression
 from black.nodes import is_string_token
 from black.lines import Line, EmptyLineTracker
@@ -450,7 +455,7 @@ def main(
         if root:
             out(
                 f"Identified `{root}` as project root containing a {method}.",
-                fg="blue",
+                style=LogLevel.info,
             )
 
             normalized = [
@@ -465,14 +470,14 @@ def main(
                     for _norm, source in normalized
                 ]
             )
-            out(f"Sources to be formatted: {srcs_string}", fg="blue")
+            out(f"Sources to be formatted: {srcs_string}", style=LogLevel.info)
 
         if config:
             config_source = ctx.get_parameter_source("config")
             if config_source in (ParameterSource.DEFAULT, ParameterSource.DEFAULT_MAP):
-                out("Using configuration from project root.", fg="blue")
+                out("Using configuration from project root.", style=LogLevel.info)
             else:
-                out(f"Using configuration in '{config}'.", fg="blue")
+                out(f"Using configuration in '{config}'.", style=LogLevel.info)
 
     error_msg = "Oh no! 💥 💔 💥"
     if (
@@ -564,7 +569,10 @@ def main(
     if verbose or not quiet:
         if code is None and (verbose or report.change_count or report.failure_count):
             out()
-        out(error_msg if report.return_code else "All done! ✨ 🍰 ✨")
+        out(
+            error_msg if report.return_code else "All done! ✨ 🍰 ✨",
+            style=LogLevel.success,
+        )
         if code is None:
             click.echo(str(report), err=True)
     ctx.exit(report.return_code)